requirements.txt
# Core Discord Bot
discord.py>=2.3.0
wavelink>=3.4.0

# Environment & Config
python-dotenv>=1.0.0

# Audio & HTTP
aiohttp>=3.8.0
aiofiles>=23.0.0

# Database
aiosqlite>=0.19.0

# Monitoring & Health
psutil>=5.9.0

# Development tools (optional)
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0